examples/requirements_lin.txt
numpy==1.19.5
opencv-python==4.5.1.48
pillow==8.1.0